Move chunker and file_walker to index
